First commit
Extracted this code from the Marketing Bridge extension. It wasn’t
dependent on anything in there so it made sense to make it it’s own
extension. This way we can use the toolbar icon to quick swap and the
original extension can now use the click functionality for a popup
window.